# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API: health, admin CRUD; route user (auth) dự phòng. Swagger: /swagger/index.html",
        "title": "TrongCon API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "TrongCon",
            "url": "https://github.com",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "Danh sách user (phân trang)",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Trang",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Số bản ghi",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_user_v1.ListUsersRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    }
                }
            },
            "post": {
                "description": "Tạo user mới (mật khẩu được hash bcrypt).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "Tạo user",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_user_v1.CreateReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_user_v1.CreateRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "Lấy user theo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_user_v1.GetRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "Cập nhật user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_user_v1.UpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_user_v1.UpdateRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "summary": "Xóa user (soft delete)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_user_v1.DeleteRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.ErrBody"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Kiểm tra API còn sống.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/trongcon-api_api_swagger.HealthOK"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "trongcon-api_api_swagger.ErrBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "trongcon-api_api_swagger.HealthOK": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "trongcon-api_api_user_v1.CreateReq": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "trongcon-api_api_user_v1.CreateRes": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/trongcon-api_api_user_v1.UserRes"
                }
            }
        },
        "trongcon-api_api_user_v1.DeleteRes": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "trongcon-api_api_user_v1.GetRes": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/trongcon-api_api_user_v1.UserRes"
                }
            }
        },
        "trongcon-api_api_user_v1.ListUsersRes": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/trongcon-api_api_user_v1.UserRes"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "trongcon-api_api_user_v1.UpdateReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "trongcon-api_api_user_v1.UpdateRes": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/trongcon-api_api_user_v1.UserRes"
                }
            }
        },
        "trongcon-api_api_user_v1.UserRes": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}